+++ index.html
@@ -231,10 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
- module4-task1
-<script src="js/main.js"></script>
-
-
- master
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
